requirements.txt
Pillow==9.3.0
pytest==7.0.1
pymd6==0.1.1